Refactoring, error handling, slack alerts, test

+++ src/main/java/com/data/collector/models/Machines.java
@@ -10,7 +10,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
 @Entity
-@Table(name="machines")
+@Table(name = "machines")
 public class Machines {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
@@ -39,7 +39,7 @@ public class Machines {
     private int gpu_ram;
 
     @Column(name = "gpu_max_cur_temp")
-    private float gpu_max_cur_temp;
+    private int gpu_max_cur_temp;
 
     @Column(name = "cpu_name")
     private String cpu_name;
@@ -50,9 +50,12 @@ public class Machines {
     @Column(name = "error_description")
     private String error_description;
 
-    public Machines() {}
+    public Machines() {
+    }
 
-    public Machines(UUID id, UUID machine_id, String hostname, int timeout, int num_gpus, float total_flops, String gpu_name, int gpu_ram, float gpu_max_cur_temp, String cpu_name, float earn_day, String error_description) {
+    public Machines(UUID id, UUID machine_id, String hostname, int timeout, int num_gpus, float total_flops,
+            String gpu_name, int gpu_ram, int gpu_max_cur_temp, String cpu_name, float earn_day,
+            String error_description) {
         this.id = id;
         this.machine_id = machine_id;
         this.hostname = hostname;
@@ -131,11 +134,11 @@ public class Machines {
         this.gpu_ram = gpu_ram;
     }
 
-    public float getGpu_max_cur_temp() {
+    public int getGpu_max_cur_temp() {
         return this.gpu_max_cur_temp;
     }
 
-    public void setGpu_max_cur_temp(float gpu_max_cur_temp) {
+    public void setGpu_max_cur_temp(int gpu_max_cur_temp) {
         this.gpu_max_cur_temp = gpu_max_cur_temp;
     }
 
@@ -165,19 +168,10 @@ public class Machines {
 
     @Override
     public String toString() {
-        return "{" +
-            " id='" + getId() + "'" +
-            ", machine_id='" + getMachine_id() + "'" +
-            ", hostname='" + getHostname() + "'" +
-            ", timeout='" + getTimeout() + "'" +
-            ", num_gpus='" + getNum_gpus() + "'" +
-            ", total_flops='" + getTotal_flops() + "'" +
-            ", gpu_name='" + getGpu_name() + "'" +
-            ", gpu_ram='" + getGpu_ram() + "'" +
-            ", gpu_max_cur_temp='" + getGpu_max_cur_temp() + "'" +
-            ", cpu_name='" + getCpu_name() + "'" +
-            ", earn_day='" + getEarn_day() + "'" +
-            ", error_description='" + getError_description() + "'" +
-            "}";
+        return "{" + " id='" + getId() + "'" + ", machine_id='" + getMachine_id() + "'" + ", hostname='" + getHostname()
+                + "'" + ", timeout='" + getTimeout() + "'" + ", num_gpus='" + getNum_gpus() + "'" + ", total_flops='"
+                + getTotal_flops() + "'" + ", gpu_name='" + getGpu_name() + "'" + ", gpu_ram='" + getGpu_ram() + "'"
+                + ", gpu_max_cur_temp='" + getGpu_max_cur_temp() + "'" + ", cpu_name='" + getCpu_name() + "'"
+                + ", earn_day='" + getEarn_day() + "'" + ", error_description='" + getError_description() + "'" + "}";
     }
 }
