+++ src/main/resources/db/migration/postgresql/V1__machine_data.sql
@@ -9,7 +9,7 @@ CREATE TABLE machines (
   total_flops FLOAT,
   gpu_name VARCHAR(255),
   gpu_ram INT,
-  gpu_max_cur_temp FLOAT,
+  gpu_max_cur_temp INT,
   cpu_name VARCHAR(255),
   earn_day FLOAT,
   error_description VARCHAR(255)
